+++ src/main/java/exercise/ExerciseManager.java
@@ -5,7 +5,7 @@ public class ExerciseManager {
     /**
      * Create new exercise instance without a type
      */
-    public Exercise createExercise(String name, int caloriesBurntPerMin) {
+    public Exercise createExercise(String name, float caloriesBurntPerMin) {
         if (name.equals("_")) {
             throw new IllegalArgumentException ("An exercise cannot be named _");
         }
@@ -15,7 +15,7 @@ public class ExerciseManager {
     /**
      * Create new exercise instance with a type
      */
-    public Exercise createExercise(String name, String type, int caloriesBurntPerMin) {
+    public Exercise createExercise(String name, String type, float caloriesBurntPerMin) {
         if (name.equals("_")) {
             throw new IllegalArgumentException ("An exercise cannot be named _");
         }
